Update Hole18FortuneTeller.ino
I've added everything we need to the loop, but the while loop is not behaving.

I added the delay(22000) at the bottom because it was throwing the specialFortune relay continuously when it was triggered rather than once per cycle.

I set it to white at each repeat of the main loop so that if it goes through the if(senseMotion) loop then it will return to white afterwards. It seems to go through the rainbow cycle once, but it doesn't trigger for senseMotion triggers after that, even though the specialFortune relay is triggering reliably.

I tried moving the initializing of the millis timer ints to before the if statement, but that didn't help.

The red light relay is on for much longer than 3000 milliseconds.